+++ src/cirq_iqm/devices/iqm_device.py
@@ -21,18 +21,19 @@ to use with the architecture.
 from __future__ import annotations
 
 import collections.abc as ca
-import uuid
 from math import pi as PI
-from typing import Optional, Union
+from typing import FrozenSet, Optional, Union
+import uuid
 
 import cirq
 from cirq import InsertStrategy, MeasurementGate, devices, ops, protocols
-from cirq.contrib.routing.router import nx, route_circuit
+from cirq.contrib.routing.router import route_circuit
+
+from .iqm_device_metadata import IQMDeviceMetadata
 
 
 def _verify_unique_measurement_keys(operations: ca.Iterable[cirq.Operation]) -> None:
-    """Raises an error if a measurement key is repeated in the given set of Operations.
-    """
+    """Raises an error if a measurement key is repeated in the given set of Operations."""
     seen_keys: set[str] = set()
     for op in operations:
         if protocols.is_measurement(op):
@@ -42,7 +43,7 @@ def _verify_unique_measurement_keys(operations: ca.Iterable[cirq.Operation]) -> 
             seen_keys.add(key)
 
 
-def _validate_for_routing(circuit: cirq.Circuit) -> None:
+def _validate_for_routing(circuit: cirq.AbstractCircuit) -> None:
     if not circuit.are_all_measurements_terminal():
         raise ValueError('Non-terminal measurements are not supported')
 
@@ -52,59 +53,41 @@ class IQMDevice(devices.Device):
 
     Adds extra functionality on top of the basic :class:`cirq.Device` class for decomposing gates,
     optimizing circuits and mapping qubits.
+
+    Args:
+        metadata: device metadata which contains the qubits, their connectivity, and the native gateset
     """
-    QUBIT_COUNT: int = None
-    """number of qubits on the device"""
 
-    QUBIT_NAME_PREFIX: str = 'QB'
-    """prefix for qubit names, to be followed by their numerical index"""
+    def __init__(self, metadata: IQMDeviceMetadata):
+        self._metadata = metadata
+        self.qubits = tuple(sorted(self._metadata.qubit_set))
 
-    CONNECTIVITY: tuple[set[int]] = ()
-    """qubit connectivity graph of the device"""
-
-    NATIVE_GATES: tuple[type[cirq.Gate]] = (
-        ops.PhasedXPowGate,
-        ops.XPowGate,
-        ops.YPowGate,
-        ops.MeasurementGate
-    )
-    """native gate set of the device (gate families)"""
-
-    NATIVE_GATE_INSTANCES: tuple[cirq.Gate] = (
-        ops.CZPowGate(),
-    )
-    """native gate set of the device (individual gates)"""
-
-    def __init__(self):
-        self.qubits = tuple(cirq.NamedQubit.range(1, self.QUBIT_COUNT + 1, prefix=self.QUBIT_NAME_PREFIX))
+    @property
+    def metadata(self) -> IQMDeviceMetadata:
+        """Returns metadata for the device."""
+        return self._metadata
 
     @classmethod
     def get_qubit_index(cls, qubit: cirq.NamedQubit) -> int:
         """The numeric index of the given qubit on the device."""
-        return int(qubit.name[len(cls.QUBIT_NAME_PREFIX):])
+        return int(qubit.name[len(IQMDeviceMetadata.QUBIT_NAME_PREFIX) :])
 
     def get_qubit(self, index: int) -> cirq.NamedQubit:
         """The device qubit corresponding to the given numeric index."""
         return self.qubits[index - 1]  # 1-based indexing
 
-    @classmethod
-    def check_qubit_connectivity(cls, operation: cirq.Operation) -> None:
-        """Raises a ValueError if operation acts on qubits that are not connected.
-        """
+    def check_qubit_connectivity(self, operation: cirq.Operation) -> None:
+        """Raises a ValueError if operation acts on qubits that are not connected."""
         if len(operation.qubits) >= 2 and not isinstance(operation.gate, ops.MeasurementGate):
-            connection = set(cls.get_qubit_index(q) for q in operation.qubits)
-            if connection not in cls.CONNECTIVITY:
+            if operation.qubits not in self._metadata.nx_graph.edges:
                 raise ValueError(f'Unsupported qubit connectivity required for {operation!r}')
 
-    @classmethod
-    def is_native_operation(cls, op: cirq.Operation) -> bool:
+    def is_native_operation(self, op: cirq.Operation) -> bool:
         """Predicate, True iff the given operation is considered native for the architecture."""
         return (
             isinstance(op, (ops.GateOperation, ops.TaggedOperation))
-            and (
-                isinstance(op.gate, cls.NATIVE_GATES)
-                or op.gate in cls.NATIVE_GATE_INSTANCES
-            )
+            and (op.gate is not None)
+            and (op.gate in self._metadata.gateset)
         )
 
     def operation_decomposer(self, op: cirq.Operation) -> Optional[list[cirq.Operation]]:
@@ -179,8 +162,7 @@ class IQMDevice(devices.Device):
         return None
 
     def decompose_operation(self, operation: cirq.Operation) -> cirq.OP_TREE:
-        """Decompose a single quantum operation into the native operation set.
-        """
+        """Decompose a single quantum operation into the native operation set."""
         if self.is_native_operation(operation):
             return operation
 
@@ -188,14 +170,11 @@ class IQMDevice(devices.Device):
             operation,
             intercepting_decomposer=self.operation_decomposer,
             keep=self.is_native_operation,
-            on_stuck_raise=None
+            on_stuck_raise=None,
         )
 
     def route_circuit(
-            self,
-            circuit: cirq.Circuit,
-            *,
-            return_swap_network: bool = False
+        self, circuit: cirq.Circuit, *, return_swap_network: bool = False
     ) -> Union[cirq.Circuit, cirq.contrib.routing.SwapNetwork]:
         """Routes the given circuit to the device connectivity.
 
@@ -215,38 +194,34 @@ class IQMDevice(devices.Device):
         """
         _validate_for_routing(circuit)
 
-        device_graph = nx.Graph(tuple(map(self.get_qubit, edge)) for edge in self.CONNECTIVITY)
-
         # Remove all measurement gates and replace them with 1-qubit identity gates so they don't
         # disappear from the final swap network if no other operations remain. We will add them back after routing the
         # rest of the network. This is done for two purposes: it allows the circuit to contain measurements of more than
         # 2 qubits in the circuit and prevents measurements becoming non-terminal during routing.
-        measurement_ops = [(m[0], m[1]) for m in circuit.findall_operations_with_gate_type(MeasurementGate)]
-        measurement_qubits = set().union(*[op.qubits for _, op in measurement_ops])
+        measurement_ops = list(circuit.findall_operations_with_gate_type(MeasurementGate))
+        measurement_qubits = set().union(*[op.qubits for _, op, _ in measurement_ops])
 
         modified_circuit = circuit.copy()
-        modified_circuit.batch_remove(measurement_ops)
+        modified_circuit.batch_remove([(ind, op) for ind, op, _ in measurement_ops])
         i_tag = uuid.uuid4()
         for q in measurement_qubits:
             modified_circuit.append(cirq.I(q).with_tags(i_tag))
 
         # Route the modified circuit.
-        swap_network = route_circuit(modified_circuit, device_graph, algo_name='greedy')
+        swap_network = route_circuit(modified_circuit, self._metadata.nx_graph, algo_name='greedy')
 
         # Return measurements to the circuit with potential qubit swaps.
         final_qubit_mapping = {v: k for k, v in swap_network.final_mapping().items()}
         new_measurements = []
-        for _, op in measurement_ops:
+        for _, op, gate in measurement_ops:
             new_qubits = [final_qubit_mapping[q] for q in op.qubits]
-            new_measurement = cirq.measure(*new_qubits, key=op.gate.key)
+            new_measurement = cirq.measure(*new_qubits, key=gate.key)
             new_measurements.append(new_measurement)
 
         swap_network.circuit.append(new_measurements, InsertStrategy.NEW_THEN_INLINE)
 
         # Remove additional identity gates.
-        identity_gates = swap_network.circuit.findall_operations(
-            lambda op: i_tag in op.tags
-        )
+        identity_gates = swap_network.circuit.findall_operations(lambda op: i_tag in op.tags)
         swap_network.circuit.batch_remove(identity_gates)
 
         if return_swap_network:
@@ -269,7 +244,7 @@ class IQMDevice(devices.Device):
         )
         return cirq.Circuit(moments)
 
-    def validate_circuit(self, circuit: cirq.Circuit) -> None:
+    def validate_circuit(self, circuit: cirq.AbstractCircuit) -> None:
         super().validate_circuit(circuit)
         _verify_unique_measurement_keys(circuit.all_operations())
         _validate_for_routing(circuit)
@@ -287,5 +262,9 @@ class IQMDevice(devices.Device):
 
         self.check_qubit_connectivity(operation)
 
+    @staticmethod
+    def _qubit_set_from_count(qubit_count: int) -> FrozenSet[cirq.NamedQubit]:
+        return frozenset(cirq.NamedQubit.range(1, qubit_count + 1, prefix=IQMDeviceMetadata.QUBIT_NAME_PREFIX))
+
     def __eq__(self, other):
-        return self.__class__ == other.__class__
+        return self.__class__ == other.__class__ and self._metadata == other._metadata