+++ tests/test_iqm_sampler.py
@@ -14,11 +14,10 @@
 import uuid
 
 import cirq
+from iqm_client.iqm_client import Circuit, IQMClient, Metadata, RunResult, Status
+from mockito import ANY, mock, when
 import pytest
 import sympy
-from iqm_client.iqm_client import (Circuit, IQMClient, Metadata, RunResult,
-                                   Status)
-from mockito import ANY, mock, when
 
 from cirq_iqm import Adonis
 from cirq_iqm.iqm_sampler import IQMSampler
@@ -28,18 +27,14 @@ from cirq_iqm.iqm_sampler import IQMSampler
 def circuit():
     qubit_1 = cirq.NamedQubit('q1 log.')
     qubit_2 = cirq.NamedQubit('q2 log.')
-    return cirq.Circuit(
-        cirq.measure(qubit_1, qubit_2, key='result')
-    )
+    return cirq.Circuit(cirq.measure(qubit_1, qubit_2, key='result'))
 
 
 @pytest.fixture()
 def circuit_with_physical_names():
     qubit_1 = cirq.NamedQubit('QB1')
     qubit_2 = cirq.NamedQubit('QB2')
-    return cirq.Circuit(
-        cirq.measure(qubit_1, qubit_2, key='result')
-    )
+    return cirq.Circuit(cirq.measure(qubit_1, qubit_2, key='result'))
 
 
 @pytest.fixture()
@@ -49,10 +44,7 @@ def iqm_metadata():
 
 @pytest.fixture()
 def qubit_mapping():
-    return {
-        'q1 log.': 'QB1',
-        'q2 log.': 'QB2'
-    }
+    return {'q1 log.': 'QB1', 'q2 log.': 'QB2'}
 
 
 @pytest.fixture()
@@ -60,21 +52,12 @@ def adonis_sampler(base_url, qubit_mapping):
     return IQMSampler(base_url, Adonis(), qubit_mapping=qubit_mapping)
 
 
-@pytest.fixture()
-def adonis_sampler_without_settings(base_url):
-    return IQMSampler(base_url, Adonis())
-
-
 @pytest.mark.usefixtures('unstub')
 def test_run_sweep_executes_circuit(adonis_sampler, circuit, iqm_metadata):
     client = mock(IQMClient)
     run_id = uuid.uuid4()
     run_result = RunResult(status=Status.READY, measurements=[{'some stuff': [[0], [1]]}], metadata=iqm_metadata)
-    when(client).submit_circuits(ANY,
-                                 qubit_mapping=ANY,
-                                 settings=ANY,
-                                 calibration_set_id=ANY,
-                                 shots=ANY).thenReturn(run_id)
+    when(client).submit_circuits(ANY, qubit_mapping=ANY, calibration_set_id=ANY, shots=ANY).thenReturn(run_id)
     when(client).wait_for_results(run_id).thenReturn(run_result)
 
     adonis_sampler._client = client
@@ -84,45 +67,33 @@ def test_run_sweep_executes_circuit(adonis_sampler, circuit, iqm_metadata):
 
 @pytest.mark.usefixtures('unstub')
 def test_run_sweep_with_bad_qubit_mapping(base_url, circuit):
-    qubit_mapping = {
-        'q1 log.': 'QB1',
-        'q2 log.': 'QB1'
-    }
+    qubit_mapping = {'q1 log.': 'QB1', 'q2 log.': 'QB1'}
     sampler = IQMSampler(base_url, Adonis(), qubit_mapping=qubit_mapping)
     with pytest.raises(ValueError, match='Failed applying qubit mapping.'):
         sampler.run_sweep(circuit, None, repetitions=2)
 
 
 @pytest.mark.usefixtures('unstub')
-def test_run_sweep_executes_circuit_without_settings(adonis_sampler_without_settings,
-                                                     circuit_with_physical_names,
-                                                     iqm_metadata):
+def test_run_sweep_executes_circuit_with_physical_names(adonis_sampler, circuit_with_physical_names, iqm_metadata):
     client = mock(IQMClient)
     run_id = uuid.uuid4()
     run_result = RunResult(status=Status.READY, measurements=[{'some stuff': [[0], [1]]}], metadata=iqm_metadata)
-    when(client).submit_circuits(ANY,
-                                 qubit_mapping=ANY,
-                                 settings=None,
-                                 calibration_set_id=ANY,
-                                 shots=ANY).thenReturn(run_id)
+    when(client).submit_circuits(ANY, qubit_mapping=ANY, calibration_set_id=ANY, shots=ANY).thenReturn(run_id)
     when(client).wait_for_results(run_id).thenReturn(run_result)
 
-    adonis_sampler_without_settings._client = client
-    results = adonis_sampler_without_settings.run_sweep(circuit_with_physical_names, None, repetitions=2)
+    adonis_sampler._client = client
+    results = adonis_sampler.run_sweep(circuit_with_physical_names, None, repetitions=2)
     assert isinstance(results[0], cirq.Result)
 
 
 @pytest.mark.usefixtures('unstub')
-def test_run_sweep_with_parameter_sweep(adonis_sampler_without_settings, iqm_metadata):
+def test_run_sweep_with_parameter_sweep(adonis_sampler, iqm_metadata):
     client = mock(IQMClient)
     run_id = uuid.uuid4()
     run_result = RunResult(
-        status=Status.READY, measurements=[{'some stuff': [[0]]}, {'some stuff': [[1]]}], metadata=iqm_metadata)
-    when(client).submit_circuits(ANY,
-                                 qubit_mapping=ANY,
-                                 settings=ANY,
-                                 calibration_set_id=ANY,
-                                 shots=ANY).thenReturn(run_id)
+        status=Status.READY, measurements=[{'some stuff': [[0]]}, {'some stuff': [[1]]}], metadata=iqm_metadata
+    )
+    when(client).submit_circuits(ANY, qubit_mapping=ANY, calibration_set_id=ANY, shots=ANY).thenReturn(run_id)
     when(client).wait_for_results(run_id).thenReturn(run_result)
     qubit_1 = cirq.NamedQubit('QB1')
     qubit_2 = cirq.NamedQubit('QB2')
@@ -131,9 +102,9 @@ def test_run_sweep_with_parameter_sweep(adonis_sampler_without_settings, iqm_met
     sweep_length = 2
     param_sweep = cirq.Linspace('t', start=0, stop=1, length=sweep_length)
 
-    adonis_sampler_without_settings._client = client
+    adonis_sampler._client = client
 
-    results = adonis_sampler_without_settings.run_sweep(circuit_sweep, param_sweep, repetitions=123)
+    results = adonis_sampler.run_sweep(circuit_sweep, param_sweep, repetitions=123)
     assert len(results) == sweep_length
     assert all(isinstance(result, cirq.Result) for result in results)