+++ tests/test_valkmusa.py
@@ -15,6 +15,7 @@
 """Tests for the Valkmusa device.
 """
 import cirq
+
 # pylint: disable=redefined-outer-name,no-self-use,duplicate-code
 import pytest
 
@@ -79,10 +80,13 @@ class TestOperationValidation:
         valkmusa.validate_operation(gate(QB1, QB2))
         valkmusa.validate_operation(gate(QB2, QB1))
 
-    @pytest.mark.parametrize('meas', [
-        cirq.measure,
-        lambda q: cirq.measure(q, key='test'),
-    ])
+    @pytest.mark.parametrize(
+        'meas',
+        [
+            cirq.measure,
+            lambda q: cirq.measure(q, key='test'),
+        ],
+    )
     def test_native_measurements(self, valkmusa, meas):
         """Native operations must pass validation."""
 
@@ -112,9 +116,12 @@ class TestOperationValidation:
         with pytest.raises(ValueError, match='Unsupported gate type'):
             valkmusa.validate_operation(gate(QB2, QB1))
 
-    @pytest.mark.parametrize('qubit', [
-        cirq.NamedQubit('xxx'),
-    ])
+    @pytest.mark.parametrize(
+        'qubit',
+        [
+            cirq.NamedQubit('xxx'),
+        ],
+    )
     def test_qubits_not_on_device(self, valkmusa, qubit):
         """Gates operating on qubits not on device must not pass validation."""
 
@@ -126,12 +133,12 @@ class TestGateDecomposition:
     """Decomposing gates."""
 
     @staticmethod
-    def is_native(op_or_op_list) -> bool:
-        """True iff the op_list consists of native operations only."""
-        if Valkmusa.is_native_operation(op_or_op_list):
+    def is_native(valkmusa, op_or_op_list) -> bool:
+        """True iff the op_list consists of native operations of valkmusa only."""
+        if valkmusa.is_native_operation(op_or_op_list):
             return True
         for op in op_or_op_list:
-            if not Valkmusa.is_native_operation(op):
+            if not valkmusa.is_native_operation(op):
                 raise TypeError(f'Non-native operation: {op}')
         return True
 
@@ -142,12 +149,12 @@ class TestGateDecomposition:
         QB1, QB2 = valkmusa.qubits
 
         for op in (
-                gate.on(QB1),
-                gate.on(QB2).with_tags('tag_baz'),
+            gate.on(QB1),
+            gate.on(QB2).with_tags('tag_baz'),
         ):
             decomposition = valkmusa.decompose_operation(op)
             assert decomposition == op
-            assert TestGateDecomposition.is_native(decomposition)
+            assert TestGateDecomposition.is_native(valkmusa, decomposition)
 
     @pytest.mark.parametrize('gate', non_native_1q_gates)
     def test_non_native_single_qubit_gates(self, valkmusa, gate):
@@ -155,11 +162,11 @@ class TestGateDecomposition:
 
         QB1, QB2 = valkmusa.qubits
         for op in (
-                gate.on(QB1),
-                gate.on(QB2).with_tags('tag_baz'),
+            gate.on(QB1),
+            gate.on(QB2).with_tags('tag_baz'),
         ):
             decomposition = valkmusa.decompose_operation(op)
-            assert TestGateDecomposition.is_native(decomposition)
+            assert TestGateDecomposition.is_native(valkmusa, decomposition)
 
     @pytest.mark.parametrize('gate', native_2q_gates)
     def test_native_two_qubit_gates(self, valkmusa, gate):
@@ -170,7 +177,7 @@ class TestGateDecomposition:
         op = gate(QB1, QB2)
         decomposition = valkmusa.decompose_operation(op)
         assert decomposition == op
-        assert TestGateDecomposition.is_native(decomposition)
+        assert TestGateDecomposition.is_native(valkmusa, decomposition)
 
     @pytest.mark.parametrize('gate', non_native_2q_gates)
     def test_non_native_two_qubit_gates(self, valkmusa, gate):
@@ -179,8 +186,8 @@ class TestGateDecomposition:
         QB1, QB2 = valkmusa.qubits
 
         for op in (
-                gate(QB1, QB2),
-                gate(QB2, QB1).with_tags('tag_baz'),
+            gate(QB1, QB2),
+            gate(QB2, QB1).with_tags('tag_baz'),
         ):
             decomposition = valkmusa.decompose_operation(op)
-            assert TestGateDecomposition.is_native(decomposition)
+            assert TestGateDecomposition.is_native(valkmusa, decomposition)