+++ CHANGELOG.rst
@@ -2,6 +2,37 @@
 Changelog
 =========
 
+Version 8.0
+===========
+
+* Implement metadata for IQM devices. `#92 <https://github.com/iqm-finland/cirq-on-iqm/pull/92>`_
+
+Version 7.8
+===========
+
+* Bump ``iqm-client`` dependency. `#91 <https://github.com/iqm-finland/cirq-on-iqm/pull/91>`_
+
+Version 7.7
+===========
+
+* Enable mypy support. `#88 <https://github.com/iqm-finland/cirq-on-iqm/pull/88>`_
+
+Version 7.6
+===========
+
+* Upgrade to IQMClient version 8.0.
+* Remove ``settings`` parameter from ``IQMSampler``.
+
+Version 7.5
+===========
+
+* Upgrade to IQMClient version 7.0.
+
+Version 7.4
+===========
+
+* ``cortex-cli`` is now the preferred way of authentication.
+
 Version 7.3
 ===========